refactor: split large files across 13 modules (Fixes #35)
Reapply v2 refactor on current main: extract helpers, hooks, and
sub-components for trip detail, AI service, chat, dashboard, store,
places, and itinerary. Add docs/codebase split plan. No behavior change.
