+++ src/features/chat/hooks/useTripChat.ts
@@ -3,7 +3,14 @@ import type { RealtimeChannel } from '@supabase/supabase-js';
 import { useTranslation } from 'react-i18next';
 import { supabase } from '@/lib/supabase';
 import { useStore } from '@/lib/store';
-import type { Message, Profile } from '@/lib/types/database.types';
+import type { Message } from '@/lib/types/database.types';
+import {
+  fetchMessagesWithProfiles,
+  sendChatMessage,
+  updateChatMessage,
+  softDeleteChatMessage,
+  type UserProfilesMap,
+} from './tripChatApi';
 
 export interface MessageWithProfile extends Message {
   sender_name?: string;
@@ -23,76 +30,25 @@ export function useTripChat({ tripId, userRole }: UseTripChatOptions) {
   const [editText, setEditText] = useState('');
   const [messageText, setMessageText] = useState('');
   const [messagesWithProfiles, setMessagesWithProfiles] = useState<MessageWithProfile[]>([]);
-  const [userProfiles, setUserProfiles] = useState<
-    Record<string, { name: string; avatar: string }>
-  >({});
+  const [, setUserProfiles] = useState<UserProfilesMap>({});
   const messagesEndRef = useRef<HTMLDivElement | null>(null);
   const channelRef = useRef<RealtimeChannel | null>(null);
 
   const user = useStore((state) => state.user);
-  const messages = useStore((state) => state.messages);
   const setMessages = useStore((state) => state.setMessages);
   const addMessage = useStore((state) => state.addMessage);
 
   const loadMessages = useCallback(async () => {
     try {
       setLoading(true);
-
-      const { data: messagesData, error: messagesError } = await supabase
-        .from('messages')
-        .select('*')
-        .eq('trip_id', tripId)
-        .is('deleted_at', null)
-        .order('created_at', { ascending: true });
-
-      if (messagesError) {
-        console.error('Error loading messages:', messagesError);
-        throw messagesError;
-      }
-
-      const mappedMessages: Message[] = (messagesData || []).map((m) => ({
-        id: m.id,
-        trip_id: m.trip_id,
-        user_id: m.user_id || '',
-        content: m.content,
-        message_type: m.message_type,
-        created_at: m.created_at,
-      }));
-
-      setMessages(mappedMessages);
-
-      const userIds = [...new Set(mappedMessages.map((m) => m.user_id).filter(Boolean))];
-      if (userIds.length > 0) {
-        const { data: profiles, error: profilesError } = await supabase
-          .from('profiles')
-          .select('id, display_name, avatar_url')
-          .in('id', userIds);
-
-        if (!profilesError && profiles) {
-          const profilesMap: Record<string, { name: string; avatar: string }> = {};
-          const rows = profiles as Pick<Profile, 'id' | 'display_name' | 'avatar_url'>[];
-          rows.forEach((profile) => {
-            profilesMap[profile.id] = {
-              name: profile.display_name || profile.id.substring(0, 8),
-              avatar:
-                profile.avatar_url ||
-                `https://api.dicebear.com/7.x/avataaars/svg?seed=${profile.id}`,
-            };
-          });
-          setUserProfiles(profilesMap);
-
-          const messagesWithProfilesData: MessageWithProfile[] = mappedMessages.map((msg) => ({
-            ...msg,
-            sender_name: msg.user_id ? profilesMap[msg.user_id]?.name : 'Unknown',
-            sender_avatar: msg.user_id ? profilesMap[msg.user_id]?.avatar : undefined,
-          }));
-          setMessagesWithProfiles(messagesWithProfilesData);
-        } else {
-          setMessagesWithProfiles(mappedMessages);
-        }
-      } else {
-        setMessagesWithProfiles(mappedMessages);
-      }
+      const {
+        messages,
+        profiles,
+        messagesWithProfiles: enriched,
+      } = await fetchMessagesWithProfiles(tripId);
+      setMessages(messages);
+      setUserProfiles(profiles);
+      setMessagesWithProfiles(enriched);
     } catch (error) {
       console.error('Error loading messages:', error);
     } finally {
@@ -105,7 +61,7 @@ export function useTripChat({ tripId, userRole }: UseTripChatOptions) {
       supabase.removeChannel(channelRef.current);
     }
 
-    const channel = supabase
+    channelRef.current = supabase
       .channel(`trip:${tripId}:messages`)
       .on(
         'postgres_changes',
@@ -120,8 +76,6 @@ export function useTripChat({ tripId, userRole }: UseTripChatOptions) {
         },
       )
       .subscribe();
-
-    channelRef.current = channel;
   }, [tripId, loadMessages]);
 
   useEffect(() => {
@@ -149,32 +103,8 @@ export function useTripChat({ tripId, userRole }: UseTripChatOptions) {
     setMessageText('');
 
     try {
-      const { data: message, error } = await supabase
-        .from('messages')
-        .insert({
-          trip_id: tripId,
-          user_id: user.id,
-          content: text,
-          message_type: 'text',
-        })
-        .select()
-        .single();
-
-      if (error) {
-        throw error;
-      }
-
-      if (message) {
-        const mappedMessage: Message = {
-          id: message.id,
-          trip_id: message.trip_id,
-          user_id: message.user_id || '',
-          content: message.content,
-          message_type: message.message_type,
-          created_at: message.created_at,
-        };
-        addMessage(mappedMessage);
-      }
+      const message = await sendChatMessage(tripId, user.id, text);
+      addMessage(message);
     } catch (error) {
       console.error('Error sending message:', error);
       setMessageText(text);
@@ -197,24 +127,10 @@ export function useTripChat({ tripId, userRole }: UseTripChatOptions) {
     if (!editText.trim() || !user) return;
 
     try {
-      const { error } = await supabase
-        .from('messages')
-        .update({
-          content: editText,
-          updated_at: new Date().toISOString(),
-        })
-        .eq('id', messageId)
-        .eq('user_id', user.id);
-
-      if (error) {
-        throw error;
-      }
-
-      const updatedMessages = messagesWithProfiles.map((msg) =>
-        msg.id === messageId ? { ...msg, content: editText } : msg,
+      await updateChatMessage(messageId, user.id, editText);
+      setMessagesWithProfiles((prev) =>
+        prev.map((msg) => (msg.id === messageId ? { ...msg, content: editText } : msg)),
       );
-      setMessagesWithProfiles(updatedMessages);
-
       setEditingMessageId(null);
       setEditText('');
     } catch (error) {
@@ -229,17 +145,7 @@ export function useTripChat({ tripId, userRole }: UseTripChatOptions) {
 
     try {
       const canDeleteAny = userRole === 'moderator' || userRole === 'owner';
-
-      const { error } = await supabase
-        .from('messages')
-        .update({ deleted_at: new Date().toISOString() })
-        .eq('id', messageId)
-        .eq(canDeleteAny ? 'trip_id' : 'user_id', canDeleteAny ? tripId : user.id);
-
-      if (error) {
-        throw error;
-      }
-
+      await softDeleteChatMessage(messageId, tripId, user.id, canDeleteAny);
       await loadMessages();
     } catch (error) {
       console.error('Error deleting message:', error);
