+++ src/lib/store/trips-slice.ts
@@ -1,6 +1,7 @@
 import type { Trip } from '../types/database.types';
 import { supabase } from '../supabase';
 import type { AppState, CreateTripData, SetState, GetState } from './types';
+import { loadTripsFromApi, createTripInApi, updateTripInApi } from './trips-api';
 
 export function createTripsSlice(set: SetState, get: GetState): Partial<AppState> {
   return {
@@ -20,61 +21,8 @@ export function createTripsSlice(set: SetState, get: GetState): Partial<AppState
 
     loadTrips: async () => {
       try {
-        const {
-          data: { user },
-        } = await supabase.auth.getUser();
-        if (!user) {
-          set({ trips: [] });
-          return;
-        }
-
-        const { data: memberships, error: membershipsError } = await supabase
-          .from('trip_members')
-          .select('trip_id')
-          .eq('user_id', user.id)
-          .is('removed_at', null);
-
-        if (membershipsError) {
-          console.error('Error loading trip memberships:', membershipsError);
-          throw membershipsError;
-        }
-
-        if (!memberships || memberships.length === 0) {
-          set({ trips: [] });
-          return;
-        }
-
-        const tripIds = (memberships as { trip_id: string }[]).map((m) => m.trip_id);
-
-        const { data: trips, error: tripsError } = await supabase
-          .from('trips')
-          .select('*')
-          .in('id', tripIds)
-          .is('deleted_at', null)
-          .order('created_at', { ascending: false });
-
-        if (tripsError) {
-          console.error('Error loading trips:', tripsError);
-          throw tripsError;
-        }
-
-        const mappedTrips: Trip[] = ((trips || []) as any[]).map((trip: any) => ({
-          id: trip.id,
-          owner_id: trip.owner_id,
-          title: trip.title,
-          destination_text: trip.destination_text,
-          start_date: trip.start_date,
-          end_date: trip.end_date,
-          status: trip.status,
-          budget_cents: trip.budget_cents ?? undefined,
-          currency: trip.currency ?? undefined,
-          constraints: trip.constraints ?? undefined,
-          active_itinerary_id: trip.active_itinerary_id ?? undefined,
-          created_at: trip.created_at,
-          updated_at: trip.updated_at,
-        }));
-
-        set({ trips: mappedTrips });
+        const trips = await loadTripsFromApi();
+        set({ trips });
       } catch (error) {
         console.error('Error loading trips:', error);
         throw error;
@@ -88,45 +36,7 @@ export function createTripsSlice(set: SetState, get: GetState): Partial<AppState
         } = await supabase.auth.getUser();
         if (!user) throw new Error('User not authenticated');
 
-        const { data: trip, error } = await supabase
-          .from('trips')
-          .insert({
-            owner_id: user.id,
-            title: tripData.title,
-            destination_text: tripData.destination_text,
-            start_date: tripData.start_date,
-            end_date: tripData.end_date,
-            status: tripData.status || 'planned',
-            budget_cents: tripData.budget_cents ?? null,
-            currency: tripData.currency ?? null,
-            constraints: tripData.constraints ?? null,
-          } as any)
-          .select()
-          .single();
-
-        if (error) {
-          console.error('Error creating trip:', error);
-          throw error;
-        }
-        if (!trip) throw new Error('Failed to create trip');
-
-        const tripFromDb = trip as any;
-        const mappedTrip: Trip = {
-          id: tripFromDb.id,
-          owner_id: tripFromDb.owner_id,
-          title: tripFromDb.title,
-          destination_text: tripFromDb.destination_text,
-          start_date: tripFromDb.start_date,
-          end_date: tripFromDb.end_date,
-          status: tripFromDb.status,
-          budget_cents: tripFromDb.budget_cents ?? undefined,
-          currency: tripFromDb.currency ?? undefined,
-          constraints: tripFromDb.constraints ?? undefined,
-          active_itinerary_id: tripFromDb.active_itinerary_id ?? undefined,
-          created_at: tripFromDb.created_at,
-          updated_at: tripFromDb.updated_at,
-        };
-
+        const mappedTrip = await createTripInApi(tripData, user.id);
         set((state) => ({ trips: [mappedTrip, ...state.trips] }));
         return mappedTrip;
       } catch (error) {
@@ -137,7 +47,7 @@ export function createTripsSlice(set: SetState, get: GetState): Partial<AppState
 
     updateTrip: async (tripId, updates) => {
       try {
-        const updateData: any = {};
+        const updateData: Record<string, unknown> = {};
         if (updates.title !== undefined) updateData.title = updates.title;
         if (updates.destination_text !== undefined)
           updateData.destination_text = updates.destination_text;
@@ -151,35 +61,7 @@ export function createTripsSlice(set: SetState, get: GetState): Partial<AppState
         if (updates.active_itinerary_id !== undefined)
           updateData.active_itinerary_id = updates.active_itinerary_id ?? null;
 
-        const { data: trip, error } = await (supabase.from('trips') as any)
-          .update(updateData)
-          .eq('id', tripId)
-          .select()
-          .single();
-
-        if (error) {
-          console.error('Error updating trip:', error);
-          throw error;
-        }
-        if (!trip) throw new Error('Trip not found');
-
-        const tripData = trip as any;
-        const mappedTrip: Trip = {
-          id: tripData.id,
-          owner_id: tripData.owner_id,
-          title: tripData.title,
-          destination_text: tripData.destination_text,
-          start_date: tripData.start_date,
-          end_date: tripData.end_date,
-          status: tripData.status,
-          budget_cents: tripData.budget_cents ?? undefined,
-          currency: tripData.currency ?? undefined,
-          constraints: tripData.constraints ?? undefined,
-          active_itinerary_id: tripData.active_itinerary_id ?? undefined,
-          created_at: tripData.created_at,
-          updated_at: tripData.updated_at,
-        };
-
+        const mappedTrip = await updateTripInApi(tripId, updateData);
         get().updateTripInState(tripId, mappedTrip);
       } catch (error) {
         console.error('Error updating trip:', error);
@@ -199,7 +81,7 @@ export function createTripsSlice(set: SetState, get: GetState): Partial<AppState
         }
 
         set((state) => ({
-          trips: state.trips.filter((t) => t.id !== tripId),
+          trips: state.trips.filter((t: Trip) => t.id !== tripId),
           currentTrip: state.currentTrip?.id === tripId ? null : state.currentTrip,
         }));
       } catch (error) {
